kafka/requests: Move partition wrapper to public header file
- This way the unit tests can use this wrapper
